Move `getBucketSize`
