+++ static/app/views/dashboards/utils/getBucketSize.tsx
